Neovim: change how types are autogenerated (#23)
Neovim is slowly starting to expose more of its api type info to lua so
we can generate more types. Generating these all in separate files is
annoying to require them all in manually (and if someone is looking to
cause trouble, could be required in without the type checker
complianing), so now this uses a 'template' file to just copy paste the
types in.